requirements.txt
agno
google-generativeai
google-genai
python-dotenv
cachetools
tqdm
python-jobspy
pandas
numpy
beautifulsoup4
requests
scikit-learn

pdfplumber
reportlab==4.0.9
xhtml2pdf==0.2.14
markdown
pypdf
rich
appwrite
Flask
flask-cors
gunicorn
tenacity
